+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.1.0 (unreleased)
 
 - **checklist**: use new GraphNG component, show units in graphs, update help texts
+- **all**: replace Angular.js based config component with React, ensure Grafana 8.0 compatibility
 - **dashboards**: PCP Vector Host Overview: add pmproxy URL and hostspec variables
 - **dashboards**: mark all dashboards as readonly
 - **vector, bpftrace**: use `pcp://127.0.0.1` as default hostspec (no functional change)